stt: move flute and pdr to stt lib
