+++ src/stt/include/stt/SteinerTreeBuilder.h
@@ -41,6 +41,7 @@
 #include <vector>
 
 #include "flute.h"
+#include "pdrev.h"
 #include "opendb/db.h"
 
 namespace ord {
