+++ src/stt/src/MakeSteinerTreeBuilder.cpp
@@ -36,7 +36,6 @@
 #include "stt/MakeSteinerTreeBuilder.h"
 
 #include "stt/SteinerTreeBuilder.h"
-#include "flute.h"
 #include "ord/OpenRoad.hh"
 #include "sta/StaMain.hh"
